requirements.txt
Django==2.0.13
django-filter==2.2.0
djangorestframework==3.10.2
Markdown==3.1.1
requests==2.22.0